# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /review-session:
    get:
      tags:
      - review_session
      summary: Get Review Session
      operationId: get_review_session_review_session_get
      parameters:
      - name: deck_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Deck Id
      - name: user_tz
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User Tz
      - name: exclude_paused
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Exclude Paused
      - name: exclude_archived
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Exclude Archived
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSessionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reviews:
    post:
      tags:
      - reviews
      summary: Create Review
      operationId: create_review_reviews_post
      parameters:
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reviews/{card_id}:
    get:
      tags:
      - reviews
      summary: Get Review History
      operationId: get_review_history_reviews__card_id__get
      parameters:
      - name: card_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Card Id
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewOut'
                title: Response Get Review History Reviews  Card Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReviewCreate:
      properties:
        card_id:
          type: string
          format: uuid
          title: Card Id
        feedback:
          $ref: '#/components/schemas/ReviewFeedback'
      type: object
      required:
      - card_id
      - feedback
      title: ReviewCreate
    ReviewFeedback:
      type: string
      enum:
      - ok
      - skipped
      - forgot
      title: ReviewFeedback
    ReviewOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        card_id:
          type: string
          format: uuid
          title: Card Id
        deck_id:
          type: string
          format: uuid
          title: Deck Id
        user_id:
          type: string
          format: uuid
          title: User Id
        reviewed_at:
          type: string
          format: date-time
          title: Reviewed At
        card_content:
          type: string
          title: Card Content
        deck_name:
          type: string
          title: Deck Name
        feedback:
          $ref: '#/components/schemas/ReviewFeedback'
        interval:
          type: integer
          title: Interval
        repetitions:
          type: integer
          title: Repetitions
        ease_factor:
          type: number
          title: Ease Factor
        created_at:
          type: string
          format: date-time
          title: Created At
        succeeded:
          type: boolean
          title: Succeeded
        failed:
          type: boolean
          title: Failed
      type: object
      required:
      - id
      - card_id
      - deck_id
      - user_id
      - reviewed_at
      - card_content
      - deck_name
      - feedback
      - interval
      - repetitions
      - ease_factor
      - created_at
      - succeeded
      - failed
      title: ReviewOut
    ReviewSessionCard:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        deck_id:
          type: string
          format: uuid
          title: Deck Id
        deck_name:
          type: string
          title: Deck Name
        content:
          type: string
          title: Content
        next_review_date:
          type: string
          format: date-time
          title: Next Review Date
        overdue:
          type: boolean
          title: Overdue
        todays_reviews:
          items:
            $ref: '#/components/schemas/ReviewOut'
          type: array
          title: Todays Reviews
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - deck_id
      - deck_name
      - content
      - next_review_date
      - overdue
      - todays_reviews
      - created_at
      - updated_at
      title: ReviewSessionCard
    ReviewSessionOut:
      properties:
        date:
          type: string
          format: date-time
          title: Date
        remaining:
          items:
            $ref: '#/components/schemas/ReviewSessionCard'
          type: array
          title: Remaining
        completed:
          items:
            $ref: '#/components/schemas/ReviewSessionCard'
          type: array
          title: Completed
        failed:
          items:
            $ref: '#/components/schemas/ReviewSessionCard'
          type: array
          title: Failed
      type: object
      required:
      - date
      - remaining
      - completed
      - failed
      title: ReviewSessionOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError